refactor: Page rearrangement
