+++ frontend/src/api/element/fixedRate.ts
