+++ frontend/src/api/element/yieldTotal.ts
