+++ frontend/src/api/element/ytcSpot.ts
