+++ frontend/src/components/Pages/YTC/Calculator/ApproveAndSimulateButton.tsx
@@ -1,17 +1,16 @@
 import { Button, ButtonProps, Spinner } from "@chakra-ui/react";
-import { BalancerApproval, ERC20Approval } from "../../../../features/approval/Approval";
+import { BalancerApproval, ERC20Approval } from "../../../Web3/Tokens/Approval";
 import { useRecoilState, useRecoilValue } from 'recoil';
 import { isSimulatedSelector, isSimulatingAtom, selectedCalculatorGainSelector, simulationResultsAtom } from "../../../../recoil/simulationResults/atom";
 import { deployments } from "../../../../constants/apy-mainnet-constants";
-import { useEffect, useState } from "react";
+import { useState } from "react";
 import { useWeb3React } from "@web3-react/core";
 import { Web3Provider } from '@ethersproject/providers';
 import { elementAddressesAtom } from "../../../../recoil/element/atom";
 import { notificationAtom } from "../../../../recoil/notifications/atom";
 import { slippageToleranceAtom } from "../../../../recoil/transactionSettings/atom";
-import { ERC20__factory } from "../../../../hardhat/typechain";
-import { getBalance } from "../../../../features/element";
-import { executeYieldTokenCompounding } from "../../../../features/ytc/executeYieldTokenCompounding";
+import { executeYieldTokenCompounding } from "../../../../api/ytc/execute";
+import { useBalance } from "../../../../hooks";
 
 interface ApproveAndSimulateButtonProps {
     tokenAddress: string | undefined;
@@ -34,7 +33,7 @@ export const ApproveAndSimulateButton: React.FC<ApproveAndSimulateButtonProps & 
                 formErrors={formErrors}
                 {...rest}
             />
-    </BalancerApproval>
+</BalancerApproval>
 }
 
 interface ApproveAndConfirmButtonProps {
@@ -44,31 +43,19 @@ export const ApproveAndConfirmButton: React.FC<ApproveAndConfirmButtonProps & Bu
     const {...rest } = props;
 
     const [isLoading, setIsLoading] = useState<boolean>(false);
-    const [balance, setBalance] = useState<number | undefined>(undefined);
     const setSimulationResults = useRecoilState(simulationResultsAtom)[1];
     const elementAddresses = useRecoilValue(elementAddressesAtom);
     const { library, account } = useWeb3React();
     const provider = library as Web3Provider;
     const setNotification = useRecoilState(notificationAtom)[1];
 
-
     const slippageTolerance = useRecoilValue(slippageToleranceAtom);
 
     const selectedResult = useRecoilValue(selectedCalculatorGainSelector);
 
+    const balance = useBalance(selectedResult?.inputs.baseTokenAddress);
+
     const sufficientBalance = selectedResult && balance && (balance >= selectedResult.inputs.amountCollateralDeposited);
-
-    // get the user balance of the token;
-    useEffect(() => {
-        if (account && selectedResult){
-            const erc20 = ERC20__factory.connect(selectedResult.inputs.baseTokenAddress, provider);
-
-            setBalance(undefined);
-            getBalance(account, erc20).then((res) => {
-                setBalance(res);
-            })
-        }
-    }, [selectedResult, account, provider])
 
     if (!selectedResult){
         return <></>
