+++ frontend/src/components/Pages/YTC/Tranche/hooks.ts
@@ -2,11 +2,11 @@ import { Web3Provider } from "@ethersproject/providers";
 import { useWeb3React } from "@web3-react/core";
 import { useCallback, useEffect } from "react";
 import { useRecoilState, useRecoilValue } from "recoil";
-import { getFixedRate } from "../../../../features/element/fixedRate";
+import { getFixedRate } from "../../../../api/element/fixedRate";
 import { elementAddressesAtom } from "../../../../recoil/element/atom";
 import { TrancheRatesInterface, trancheSelector } from "../../../../recoil/trancheRates/atom";
 import { useTokenName, useVariableAPY } from "../Calculator/hooks";
-import { yieldTokenAccruedValue } from "../../../../features/ytc/ytcHelpers";
+import { yieldTokenAccruedValue } from "../../../../api/ytc/helpers";
 import { Tranche } from "../../../../types/manual/types";
 import { getRemainingTrancheYears, getTrancheByAddress } from "../../../../features/element";
