+++ frontend/src/components/Web3/Tokens/Approval.tsx
@@ -1,17 +1,16 @@
 import React, { ReactElement, useCallback, useEffect, useState } from 'react'
 import { Button, ButtonProps, Spinner } from '@chakra-ui/react'
-import { checkApproval, sendApproval } from './approvalAPI';
+import { checkApproval, sendApproval } from '../../../api/approvalAPI';
 import { BigNumber, ContractReceipt, utils, providers } from 'ethers';
-import { notificationAtom } from '../../recoil/notifications/atom';
+import { notificationAtom } from '../../../recoil/notifications/atom';
 import { useRecoilState } from 'recoil';
-import { deployments } from '../../constants/apy-mainnet-constants';
+import { deployments } from '../../../constants/apy-mainnet-constants';
 import { useWeb3React } from '@web3-react/core';
-import { injected } from '../../connectors';
+import { injected } from '../../../connectors';
 import { Web3Provider } from '@ethersproject/providers';
-import { ERC20__factory } from '../../hardhat/typechain/factories/ERC20__factory';
-import { YieldTokenCompounding__factory } from '../../hardhat/typechain/factories/YieldTokenCompounding__factory';
-
-const MAX_UINT_HEX = '0xffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffff';
+import { ERC20__factory } from '../../../hardhat/typechain/factories/ERC20__factory';
+import { YieldTokenCompounding__factory } from '../../../hardhat/typechain/factories/YieldTokenCompounding__factory';
+import { MAX_UINT_HEX } from '../../../constants/static';
 
 type AbstractApprovalProps = {
     approveText: string,
