+++ frontend/src/constants/static.ts
@@ -1,0 +1,7 @@
+export const GITHUB_REPO = "https://github.com/Component-fi/YieldTokenCompounding";
+export const MEDIUM = "https://medium.com/@component_general";
+export const TWITTER = "https://twitter.com/componentfi";
+
+export const MAX_UINT_HEX = '0xffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffff';
+export const ZERO_ADDRESS = "0x0000000000000000000000000000000000000000";
+export const BURN_ADDRESS = "0x000000000000000000000000000000000000dead";
